requirements.txt
absl-py==1.0.0
adal==1.2.7
adlfs==2022.7.0
aiobotocore==2.3.4
aiohttp==3.8.1
aiohttp-retry==2.6.1
aioitertools==0.10.0
aiosignal==1.2.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.1
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
asyncssh==2.11.0
atomicwrites==1.4.0
atpublic==3.0.1
attrs==21.4.0
Automat==20.2.0
azure-core==1.25.0
azure-datalake-store==0.0.52
azure-identity==1.10.0
azure-storage-blob==12.13.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.10.0
#bertopic==0.9.4
billiard==3.6.4.0
bleach==5.0.0
blis==0.7.5
bokeh==2.4.2
boto3==1.21.21
botocore==1.24.21
bs4==0.0.1
cachetools==5.0.0
catalogue==2.0.6
celery==5.2.7
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
configparser==5.2.0
constantly==15.1.0
coverage==6.3.2
coveralls==3.3.1
crcmod==1.7
cryptography==36.0.2
cycler==0.11.0
cymem==2.0.6
Cython==0.29.26
dask==2021.12.0
dataprep==0.4.3
dateparser==1.1.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
dnspython==2.2.1
docopt==0.6.2
dpath==2.0.6
dulwich==0.20.45
dvc==2.20.1
dvc-data==0.1.22
dvc-http==0.0.2
dvc-objects==0.1.16
dvc-render==0.0.9
dvc-task==0.1.2
dvclive==0.10.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
facebook-sdk==3.1.0
fastjsonschema==2.15.3
filelock==3.4.2
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==2.0
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.28.4
frozenlist==1.3.0
fsspec==2022.7.1
ftfy==6.1.1
funcy==1.17
future==0.18.2
gast==0.5.3
gcsfs==2022.7.1
gensim==3.8.3
genson==1.2.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grandalf==0.6
greenlet==1.1.2
grpcio==1.44.0
h11==0.12.0
h5py==3.6.0
hdbscan==0.8.27
hnswlib==0.6.2
html5lib==1.1
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
huggingface-hub==0.4.0
humanize==4.1.0
hydra-core==1.2.0
hyperlink==21.0.0
hypothesis==6.45.0
idna==3.2
imbalanced-learn==0.9.0
imblearn==0.0
importlib-metadata==4.11.1
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.0.1
jsonpath-ng==1.5.3
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kafka-python==2.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
knack==0.9.0
kombu==5.2.4
langcodes==3.3.0
libclang==13.0.0
llvmlite==0.36.0
locket==1.0.0
lxml==4.9.1
mailchecker==4.1.19
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
maya==0.6.1
Metaphone==0.6
mistune==0.8.4
mock==4.0.3
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.2
murmurhash==1.0.6
mysql-connector==2.2.9
mysql-connector-python==8.0.26
#mysqlclient==2.0.3
nanotime==0.5.2
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
nltk==3.7
notebook==6.4.11
numba==0.53.1
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.2.0
omegaconf==2.2.3
openpyxl==3.0.9
opt-einsum==3.3.0
oss2==2.16.0
ossfs==2021.8.0
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
pathy==0.6.1
patsy==0.5.1
pendulum==2.1.2
phonenumbers==8.12.53
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.5.0
pluggy==1.0.0
ply==3.11
portalocker==2.5.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.18.0
psutil==5.9.0
psycopg2-binary==2.9.3
pure-eval==0.2.2
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.8.2
pydot==1.4.2
PyDrive2==1.14.0
pygit2==1.10.0
Pygments==2.12.0
pygtrie==2.5.0
PyJWT==2.4.0
pymongo==4.0.1
PyMySQL==1.0.2
pynndescent==0.5.6
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.1
pystache==0.6.0
pytest==7.1.2
pytest-cov==3.0.0
python-benedict==0.25.2
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-fsutil==0.6.1
python-Levenshtein==0.12.2
python-slugify==6.1.2
python-stdnum==1.17
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
#pywin32==304
#pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
regex==2021.10.23
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3fs==2022.7.1
s3transfer==0.5.2
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.1
scmrepo==0.0.25
seaborn==0.12.0
Send2Trash==1.8.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
shortuuid==1.0.9
shtab==1.5.5
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
snaptime==0.2.4
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.36
srsly==2.4.2
sshfs==2022.6.0
stack-data==0.2.0
statsmodels==0.12.2
tabpy==2.5.0
tabulate==0.8.10
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-text==2.8.1
termcolor==1.1.0
terminado==0.13.3
text-unidecode==1.3
textblob==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
threadpoolctl==2.2.0
tinycss2==1.1.1
tmdbv3api==1.7.6
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.11.2
top2vec==1.0.26
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.15.0
tweepy==4.8.0
tweet-preprocessor==0.6.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typer==0.4.0
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
umap-learn==0.5.2
uritemplate==4.1.1
urllib3==1.26.6
varname==0.8.3
vine==5.0.0
voluptuous==0.13.1
wasabi==0.9.0
wcwidth==0.2.5
webdav4==0.9.7
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wordcloud==1.8.1
wrapt==1.13.3
xmltodict==0.13.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.7.0
zope.interface==5.4.0
